refactor!: remove `Serve` namespace in `aria2_server.server`
