+++ src/aria2_server/app/core/_tools/__init__.py
@@ -1,4 +1,3 @@
-from aria2_server.app.core._tools._make_args_required import make_args_required
 from aria2_server.app.core._tools._namepace_mixin import NamepaceMixin
 
-__all__ = ("NamepaceMixin", "make_args_required")
+__all__ = ("NamepaceMixin",)
