+++ src/aria2_server/app/server/utils.py
@@ -1,0 +1,3 @@
+from aria2_server.app.core._server_config import get_server_config_in_db
+
+__all__ = ("get_server_config_in_db",)
